pyproject.toml
[tool.poetry]
name = "azcam-itl"
version = "22.2"
description = "Azcam extension for ITL systems"
authors = ["Michael Lesser <mlesser@arizona.edu>"]
license = "LICENSE"
readme = "README.md"
keywords=["ccd", "imaging", "astronomy", "sensors"]
repository = "https://github.com/mplesser/azcam-itl/"
documentation = "https://mplesser.github.io/azcam/"
include = ["status/**/*"]

[tool.poetry.dependencies]
python = "^3.9,<3.11"
numpy = "*"
ipython = "*"
rich = "*"
matplotlib = "*"
keyring = "*"
pyserial = "*"
pyvisa = "*"
mysql-connector-python = "*"
opencv-python = "*"
scikit-image = "*"
azcam = { git = "https://github.com/mplesser/azcam.git", branch = "main" }
azcam-monitor = { git = "https://github.com/mplesser/azcam-monitor.git", branch = "main" }
azcam-fastapi = { git = "https://github.com/mplesser/azcam-fastapi.git", branch = "main" }
azcam-arc = { git = "https://github.com/mplesser/azcam-arc.git", branch = "main" }
azcam-archon = { git = "https://github.com/mplesser/azcam-archon.git", branch = "main" }
azcam-ascom = { git = "https://github.com/mplesser/azcam-ascom.git", branch = "main" }
azcam-cryocon = { git = "https://github.com/mplesser/azcam-cryocon.git", branch = "main" }
azcam-ds9 = { git = "https://github.com/mplesser/azcam-ds9.git", branch = "main" }
azcam-expstatus = { git = "https://github.com/mplesser/azcam-expstatus.git", branch = "main" }
azcam-focus = { git = "https://github.com/mplesser/azcam-focus.git", branch = "main" }
azcam-imageserver = { git = "https://github.com/mplesser/azcam-imageserver.git", branch = "main" }
azcam-mag = { git = "https://github.com/mplesser/azcam-mag.git", branch = "main" }
azcam-observe = { git = "https://github.com/mplesser/azcam-observe.git", branch = "main" }
azcam-scripts = { git = "https://github.com/mplesser/azcam-scripts.git", branch = "main" }
azcam-testers = { git = "https://github.com/mplesser/azcam-testers.git", branch = "main" }
azcam-webtools = { git = "https://github.com/mplesser/azcam-webtools.git", branch = "main" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
